backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Website Cloner API - Clone websites using AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "uvicorn[standard]>=0.24.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "pillow>=10.1.0",
    "openai>=1.12.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "cssutils>=2.9.0",
    "anthropic>=0.7.8",
    "supabase>=2.3.0",
    "browserbase>=1.4.0",
]
